+++ packages/components/CHANGELOG.md
@@ -24,6 +24,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   ([#2872](https://github.com/porsche-design-system/porsche-design-system/pull/2872))
 - Vue: `theme: 'light' | 'dark' | 'auto'` prop to `PorscheDesignSystemProvider` to set `theme` on all child components  
   ([#2872](https://github.com/porsche-design-system/porsche-design-system/pull/2872))
+- Validation for usage of different PDS versions
+  ([#2867](https://github.com/porsche-design-system/porsche-design-system/pull/2867))
 
 #### Changed